move faucet related code to faucet tool folder
